+++ src/components/hero/index.ts
@@ -1,0 +1,5 @@
+export { EntityHeroLayout } from "./EntityHeroLayout";
+export { EntityHero } from "./EntityHero";
+export { EntityHeroAvatar } from "./EntityHeroAvatar";
+export { EntityHeroMetaRow } from "./EntityHeroMetaRow";
+export { EntitySection } from "./EntitySection";